Move errors from lib to a separate module
